Update packagist name and convert to vendormodule
The packagist name for this repository seems to have changed. If you google "silverstripe maintenance mode" and click on the addons page, it points at the following URLs:
Addons:
    https://addons.silverstripe.org/add-ons/thisisbd/silverstripe-maintenance-mode
Homepage:
    http://github.com/dljoseph/silverstripe-maintenance-mode 
Packagist:
    https://packagist.org/packages/thisisbd/silverstripe-maintenance-mode
Repository:
    https://github.com/dljoseph/silverstripe-maintenance-mode

Changing the package name allows the module to be installed via composer. Changing it to vendormodule allows it to be installed in the vendors folder (see issue #31 )